github-ci: Handle switch to lua by default
Ticket: #4776

+++ src/detect-lua-extensions.c
@@ -160,7 +160,7 @@ static int GetFlowVarByKey(lua_State *luastate, Flow *f, FlowVar **ret_fv)
         LUA_ERROR("key len out of range: max 256");
     }
 
-    FlowVar *fv = FlowVarGetByKey(f, (const uint8_t *)keystr, keylen);
+    FlowVar *fv = FlowVarGetByKey(f, (const uint8_t *)keystr, (uint16_t)keylen);
     if (fv == NULL) {
         LUA_ERROR("no flow var");
     }
@@ -272,7 +272,7 @@ static int LuaSetFlowvarById(lua_State *luastate)
     memcpy(buffer, str, len);
     buffer[len] = '\0';
 
-    FlowVarAddIdValue(f, idx, buffer, len);
+    FlowVarAddIdValue(f, idx, buffer, (uint16_t)len);
     return 0;
 }
 
@@ -333,7 +333,7 @@ static int LuaSetFlowvarByKey(lua_State *luastate)
     }
     memcpy(keybuf, keystr, keylen);
     keybuf[keylen] = '\0';
-    FlowVarAddKeyValue(f, keybuf, keylen, buffer, len);
+    FlowVarAddKeyValue(f, keybuf, (uint16_t)keylen, buffer, (uint16_t)len);
 
     return 0;
 }
